Fix error, add access to stat by role

+++ mail-back/mail-back/MailSender/MailSender.cs
@@ -1,4 +1,6 @@
-﻿using Microsoft.Extensions.Configuration;
+﻿using mail_back.Models;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.Options;
 using NLog;
 using System;
 using System.Collections.Generic;
@@ -13,44 +15,51 @@ namespace mail_back.Mail
     public class MailSender: IMailSender
     {
         Logger logger = LogManager.GetCurrentClassLogger();
-        string loginMail;
-        string mailPassword;
-        string smtpName;
-        int smtpPort;
-        public MailSender(IConfiguration configuration)
+
+        private MailSenderConfig mailConfug;
+        private const string displayMailName = "TaskMail";
+        private const string subjectMail = "Данные от API";
+        private string messageBody = $"<h1>Данные на момент времени {DateTime.Now}</h1>";
+
+        public MailSender(IOptionsSnapshot<MailSenderConfig> mailConfug)
         {
-            loginMail = configuration.GetSection("MailSender").GetSection("login").Value;
-            mailPassword = configuration.GetSection("MailSender").GetSection("password").Value;
-            smtpName = configuration.GetSection("SMTP").GetSection("name").Value;
-            smtpPort = Convert.ToInt32(configuration.GetSection("SMTP").GetSection("port").Value);
+            this.mailConfug = mailConfug.Value;
         }
 
         // Принимает имя файла и отправляет этот файл в письме
-        public async Task Send(string fileName, string userMail)
+        public async System.Threading.Tasks.Task Send(string fileName, string userMail)
         {
             try
             {
-                MailAddress from = new MailAddress(loginMail, "TaskMail");
-                MailAddress to = new MailAddress(userMail);
-                MailMessage mail = new MailMessage(from, to);
-                mail.Subject = "Task";
-                mail.Body = "<h1>Тест</h1>";
-                mail.IsBodyHtml = true;
-                mail.Attachments.Add(new Attachment(fileName));
-                using (SmtpClient smtp = new SmtpClient(smtpName, smtpPort))
+                if (!string.IsNullOrWhiteSpace(fileName) && !string.IsNullOrWhiteSpace(userMail))
                 {
-                    smtp.Credentials = new NetworkCredential(loginMail, mailPassword);
-                    smtp.EnableSsl = true;
-                    await smtp.SendMailAsync(mail);
-                    smtp.Dispose();
+                    MailAddress from = new MailAddress(mailConfug.Login, displayMailName);
+                    MailAddress to = new MailAddress(userMail);
+                    MailMessage mail = new MailMessage(from, to);
+                    mail.Subject = subjectMail;
+                    mail.Body = messageBody;
+                    mail.IsBodyHtml = true;
+                    mail.Attachments.Add(new Attachment(fileName));
+                    using (SmtpClient smtp = new SmtpClient(mailConfug.SmtpName, mailConfug.Port))
+                    {
+                        smtp.Credentials = new NetworkCredential(mailConfug.Login, mailConfug.Password);
+                        smtp.EnableSsl = true;
+                        await smtp.SendMailAsync(mail);
+                        smtp.Dispose();
+                    }
+                    mail.Dispose();
+                    DeleteFile(fileName);
                 }
-                mail.Dispose();
-                DeleteFile(fileName);
+                else
+                {
+                    throw new ArgumentException("fileName, userMail");
+                }
             }
-            catch(Exception ex)
+            catch (Exception ex)
             {
                 logger.Error(ex.Message);
             }
+            
         }
 
         public void DeleteFile(string fileName)
